move `call_local` into a method of `ValueInternal`, so there's less of a weird inconsistency between the callee being something other than a Handle vs the other args
