+++ crates/neon/src/types_impl/private.rs
@@ -1,9 +1,32 @@
+use std::{ffi::c_void, mem::MaybeUninit};
+
 use crate::{
-    context::internal::Env,
+    context::{internal::Env, Context},
     handle::{internal::TransparentNoCopyWrapper, Handle},
-    sys::raw,
+    result::{JsResult, NeonResult, Throw},
+    sys::{self, bindings as napi, raw},
     types::Value,
 };
+
+use super::JsValue;
+
+// Maximum number of function arguments in V8.
+const V8_ARGC_LIMIT: usize = 65535;
+
+pub(crate) unsafe fn prepare_call<'a, 'b, C: Context<'a>>(
+    cx: &mut C,
+    args: &[Handle<'b, JsValue>],
+) -> NeonResult<(usize, *const c_void)> {
+    // Note: This cast is only save because `Handle<'_, JsValue>` is
+    // guaranteed to have the same layout as a pointer because `Handle`
+    // and `JsValue` are both `repr(C)` newtypes.
+    let argv = args.as_ptr().cast();
+    let argc = args.len();
+    if argc > V8_ARGC_LIMIT {
+        return cx.throw_range_error("too many arguments");
+    }
+    Ok((argc, argv))
+}
 
 pub trait ValueInternal: TransparentNoCopyWrapper + 'static {
     fn name() -> &'static str;
@@ -29,4 +52,46 @@ pub trait ValueInternal: TransparentNoCopyWrapper + 'static {
     // # Safety
     // JavaScript value must be of type `Self`
     unsafe fn from_local(env: Env, h: raw::Local) -> Self;
+
+    unsafe fn try_call<'a, 'b, C: Context<'a>, T, AS>(
+        &self,
+        cx: &mut C,
+        this: Handle<'b, T>,
+        args: AS,
+    ) -> JsResult<'a, JsValue>
+    where
+        T: Value,
+        AS: AsRef<[Handle<'b, JsValue>]>,
+    {
+        let callee = self.to_local();
+        let (argc, argv) = unsafe { prepare_call(cx, args.as_ref()) }?;
+        let env = cx.env();
+        let mut result: MaybeUninit<raw::Local> = MaybeUninit::zeroed();
+
+        let status = napi::call_function(
+            env.to_raw(),
+            this.to_local(),
+            callee,
+            argc,
+            argv.cast(),
+            result.as_mut_ptr(),
+        );
+
+        match status {
+            sys::Status::InvalidArg if !sys::tag::is_function(env.to_raw(), callee) => {
+                return cx.throw_error("not a function");
+            }
+            sys::Status::PendingException => {
+                return Err(Throw::new());
+            }
+            status => {
+                assert_eq!(status, sys::Status::Ok);
+            }
+        }
+
+        Ok(Handle::new_internal(JsValue::from_local(
+            env,
+            result.assume_init(),
+        )))
+    }
 }
